cheap hack to get rid of spacing issues
Previously, the "<space>(link)" often meant that there would be an
obvious large blank space between the header and the following text.
Only when you hovered over the title would the "(link)" popup to fill
that space.

This is a cheap hack to ensure the header text wraps to the next line
if necessary.

A better approach would be to have the "(link)" hover out to the
margin and not interfere with spacing of the header text.